// src/app/models/jobMaster.model.ts
export class JobMaster{
  id?: number;
  order_details_id: number;
  date: string;
  date_of_delivery?: string;
  gross_weight: number;
  karigarh_id: number;
  approx_gold?: number;
  p_loss?: number;
  size?: string;
  price?: number;
  return_quantity?: number;
  material_name?: string;
  material_id?: number;
  job_number: string;
  quantity?: number;
  order_number?: string;
  model_number?: string;
  constructor() {
  }
}

// export class CustomerModel {
//   mobile1: string;
//   mobile2: string;
//   address1: string;
//   address2: string;
//   city: string;
//   state: string;
//   po: string;
//   area: string;
//   pin: string;
//   constructor(public id: number,
//               public person_name: string,
//               public email: string,
//               public customer_category_id: number
//   ) {}
// }